style: 견적조회, 반려요청 페이지 반응형 조정
